fix makefiles
