+++ email-notif-service/Makefile
@@ -1,5 +1,5 @@
 protoc:
 	protoc subs-payment.proto --proto_path ../subs-payment-service/proto --go_out . --go-grpc_out .
 	protoc email-notif.proto --proto_path proto --go_out . --go-grpc_out .
-	protoc user.proto --proto_path ../user-service/pb --go_out pb --go-grpc_out pb
-	protoc air_quality.proto --proto_path ../air-quality-service/pb --go_out pb --go-grpc_out pb
+	protoc --proto_path=../user-service/pb user.proto --go_out=paths=source_relative:pb --go-grpc_out=paths=source_relative:pb
+	protoc --proto_path=../air-quality-service/pb air_quality.proto --go_out=paths=source_relative:pb --go-grpc_out=paths=source_relative:pb
