--- subs-payment-service/util/util.go
@@ -1,1 +1,0 @@
-package util
